requirements.txt
pyirsdk>=1.3.5
pyttsx3>=2.90
pandas>=2.2.0
numpy>=1.26.0
pyinstaller>=6.0.0
